Overwrite challenge config
